// deploy/restpe-contabilidad-front-end/src/app/modules/contabilidad/domain/models/detraccion.entity.ts
export interface DetraccionEntity {
  detraccion_codigo: string;
  detraccion_codigo_sun?: string;
  detraccion_nombre_op?: string;
  detraccion_codigo_anexo?: string;
  detraccion_retencion_de_ley?: string;
  detraccion_nombre_retencion?: string;
  detraccion_porcentaje: number;
  detraccion_cuenta_c: string;
  detraccion_vigente: string;
  detraccion_vigencia_h?: string;
  detraccion_observacion?: string;
  detraccion_estado: string;
  detraccion_fecha_creacion?: string;
}